# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Batmeori API
  version: 2.0.0
  description: >-
    Provider-neutral P0 REST contract. New P0 publishes use structure-v2 and
    ontology-v2 for onion and strawberry, vi/ne, quantity
    changes, CO_PRESENT owner briefing, REMOTE anonymous language links, and
    a 24-hour anonymous today-team QR roster for person-specific assignments.
    Owner mutations require a signed workspace/team cookie and exact allowed Origin, except the explicitly public start and credential endpoints;
    worker links require neither. Clients never submit owner_id or farm_id.
    Issued worker URLs are browser /w/{token} routes.
servers:
  - url: https://api.example.invalid
    description: Deployment value is supplied by environment.
...
paths:
  /api/v1/owner/start:
    post:
      tags: [Owner]
      summary: Start or resume an input-free temporary team workspace
      description: Exact Origin and rate limit required. Random UUID Idempotency-Key reuses pending workspace on retry. Pending expires after one hour; first atomic publish activates the team's fixed 24 hours. An existing valid temporary-team cookie resumes its team; a legacy farm cookie starts a new temporary workspace.
      security: []
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses:
        '201': { description: Workspace cookie issued., content: { application/json: { schema: { $ref: '#/components/schemas/OwnerSession' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ProviderUnavailable' }
  /api/v1/owner/team-session:
    post:
      tags: [Owner]
      summary: Resume an active temporary team with its management link and PIN
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TeamPinLoginRequest' } } }
      responses:
        '201': { description: Team-scoped secure cookie issued., content: { application/json: { schema: { $ref: '#/components/schemas/OwnerSession' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ProviderUnavailable' }
  /api/v1/work-team-members/me/assignments/{sessionId}/acknowledgement:
    post:
      tags: [Worker, Team]
      summary: Explicitly acknowledge the currently displayed assigned version
      description: Exact Origin required. Checks active member assignment and team expiry. Same-version retry preserves first acknowledged_at; stale version is 409. Reading alone never acknowledges.
      security: [{ teamMemberSession: [] }]
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, additionalProperties: false, required: [expected_version], properties: { expected_version: { type: integer, minimum: 1 } } } } }
      responses:
        '200': { description: Recorded acknowledgement., content: { application/json: { schema: { $ref: '#/components/schemas/AssignmentReceipt' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/VersionConflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /api/v1/owner/session:
    post:
      tags: [Owner]
      summary: Legacy provisioned-farm authentication compatibility endpoint
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PinLoginRequest' }
      responses:
        '201':
          description: Session issued.
          headers:
            Set-Cookie:
              schema: { type: string }
              description: HttpOnly; Secure; SameSite=None cookie. Later owner mutations also require exact allowed Origin.
          content: { application/json: { schema: { $ref: '#/components/schemas/OwnerSession' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ProviderUnavailable' }
    get:
      tags: [Owner]
      summary: Read the current owner session and farm
      security: [{ ownerSession: [] }]
      responses:
        '200': { description: Current farm-scoped owner session., content: { application/json: { schema: { $ref: '#/components/schemas/OwnerSession' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [Owner]
      summary: Clear the current owner session cookie
      security: []
      responses:
        '204': { description: Session cookie cleared. }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/work-sessions:
    get:
      tags: [Owner]
      summary: List owner work sessions
      security: [{ ownerSession: [] }]
      responses:
        '200': { description: Sessions., content: { application/json: { schema: { $ref: '#/components/schemas/OwnerWorkSessionList' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/v1/work-teams/today:
    post:
      tags: [Owner, Team]
      summary: Create or return today's anonymous QR invite
      description: Temporary owners receive their active team regardless of calendar date; pending is 409. Legacy owners retain the farm/day creation path. Repeated calls return the same QR, reconstructed from server-only secret material.
      security: [{ ownerSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses:
        '200': { description: Existing team with the same active QR., content: { application/json: { schema: { $ref: '#/components/schemas/TodayWorkTeam' } } } }
        '201': { description: New team with QR., content: { application/json: { schema: { $ref: '#/components/schemas/TodayWorkTeam' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '503': { $ref: '#/components/responses/ProviderUnavailable' }
    get:
      tags: [Owner, Team]
      summary: Read today's temporary roster
      security: [{ ownerSession: [] }]
      responses:
        '200': { description: Today's roster and the same active QR URL., content: { application/json: { schema: { $ref: '#/components/schemas/TodayWorkTeam' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/work-team-members/me/assignments:
    get:
      tags: [Worker, Team]
      summary: Read this browser's assigned latest work states
      security: [{ teamMemberSession: [] }]
      responses:
        '200': { description: Latest states; an empty array means waiting for owner assignment., content: { application/json: { schema: { $ref: '#/components/schemas/TeamAssignmentsResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    ownerSession:
      type: apiKey
      in: cookie
      name: batmeori_owner_session
    teamMemberSession:
      type: apiKey
      in: cookie
      name: batmeori_team_member
  parameters:
...
    IdempotencyKey: { in: header, name: Idempotency-Key, required: true, schema: { type: string, minLength: 8 } }
  responses:
    Unauthorized: { description: Authentication required., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Forbidden: { description: Origin is not allowed., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: Resource not found., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    VersionConflict: { description: Expected version mismatch., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unprocessable: { description: Schema or safety gate rejected request., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    ProviderUnavailable: { description: AI/provider or dependency unavailable., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    TooManyRequests: { description: Rate limited., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, enum: [UNAUTHORIZED, NOT_FOUND, ACCESS_DENIED, LINK_EXPIRED, VERSION_CONFLICT, LEGACY_READ_ONLY, OVERRIDE_NOT_ALLOWED, AUDIO_UNCLEAR, SCHEMA_INVALID, PROVIDER_UNAVAILABLE, RATE_LIMITED, INTERNAL_ERROR] }
        message: { type: string }
        request_id: { type: string }
        details: { type: object, additionalProperties: true }
    PinLoginRequest:
      type: object
      additionalProperties: false
      required: [farm_code, pin]
      properties:
        farm_code: { type: string, minLength: 3, maxLength: 32, pattern: '^[a-z0-9][a-z0-9-]{2,31}$' }
        pin: { type: string, minLength: 4, maxLength: 32, writeOnly: true }
    OwnerFarm:
      type: object
      additionalProperties: false
      required: [code, display_name]
      properties:
        code: { type: string }
        display_name: { type: string }
    OwnerSession:
      type: object
      additionalProperties: false
      required: [authenticated, expires_at, farm]
      properties:
        authenticated: { type: boolean, const: true }
        expires_at: { type: string, format: date-time }
        farm: { $ref: '#/components/schemas/OwnerFarm' }
        team: { anyOf: [{ $ref: '#/components/schemas/OwnerTeamAccess' }, { type: 'null' }] }
    TeamPinLoginRequest:
      type: object
      additionalProperties: false
      required: [team_id, pin]
      properties:
        team_id: { type: string, format: uuid }
        pin: { type: string, pattern: '^[0-9]{6}$', writeOnly: true }
    OwnerTeamAccess:
      type: object
      additionalProperties: false
      required: [team_id, status, expires_at, management_url, pin]
      properties:
        team_id: { type: string, format: uuid }
        status: { type: string, enum: [PENDING, ACTIVE] }
        expires_at: { type: string, format: date-time }
        management_url: { type: [string, 'null'], format: uri }
        pin: { type: [string, 'null'], description: Owner-only management PIN. Never included in worker QR or briefing. }
    AssignmentReceipt:
      type: object
      additionalProperties: false
      required: [work_session_id, current_version, acknowledged_version, acknowledged_at]
      properties:
        work_session_id: { type: string, format: uuid }
        current_version: { type: integer, minimum: 1 }
        acknowledged_version: { type: [integer, 'null'], minimum: 1 }
        acknowledged_at: { type: [string, 'null'], format: date-time }
    TeamMember:
      type: object
      additionalProperties: false
      required: [member_id, display_name, language_code, joined_at, assignment_session_ids]
      properties:
        member_id: { type: string, format: uuid }
        display_name: { type: string, minLength: 1, maxLength: 30 }
        language_code: { type: string, enum: [vi, ne] }
        joined_at: { type: string, format: date-time }
        assignment_session_ids: { type: array, uniqueItems: true, items: { type: string, format: uuid } }
        assignment_receipts: { type: array, items: { $ref: '#/components/schemas/AssignmentReceipt' } }
    TodayWorkTeam:
      type: object
      additionalProperties: false
      required: [team_id, work_date, status, join_url, expires_at, members]
      properties:
        team_id: { type: string, format: uuid }
        work_date: { type: string, format: date, description: Asia/Seoul work date. }
        status: { type: string, const: ACTIVE }
        join_url: { type: string, format: uri, description: Stable for this team until the owner explicitly rotates it; rotation does not extend expiry. }
        expires_at: { type: string, format: date-time }
        members: { type: array, items: { $ref: '#/components/schemas/TeamMember' } }
...
    TeamAssignmentsResponse:
      type: object
      additionalProperties: false
      required: [assignments, receipts]
      properties:
        assignments: { type: array, items: { $ref: '#/components/schemas/WorkerBriefing' } }
        receipts: { type: array, items: { $ref: '#/components/schemas/AssignmentReceipt' } }
    Quantity:
      type: object
      required: [value, unit]
      properties: { value: { type: integer, minimum: 1 }, unit: { type: string, minLength: 1 } }
    DeicticLocation:
      type: object
      required: [raw_text, kind, canonical_name]
      properties:
        raw_text: { type: [string, 'null'] }
        kind: { type: string, enum: [DEICTIC, NAMED, UNSPECIFIED] }
        canonical_name: { type: [string, 'null'] }
...
    RiskAssessment:
      type: object
      required: [level, reasons, schema_version, contract_version]
      properties:
        level: { type: string, enum: [LOW, HIGH, UNKNOWN] }
        reasons: { type: array, uniqueItems: true, items: { type: string, enum: [VEHICLE_OPERATION, ROTATING_BLADE, PESTICIDE_OR_CHEMICAL, WORK_AT_HEIGHT, POWERED_MACHINERY, INSUFFICIENT_CONTEXT, OTHER_HIGH_RISK] } }
        schema_version: { type: string, const: '1' }
        contract_version: { type: string, const: safety-policy-v1 }
    Step:
      type: object
      required: [sequence, task_code, title_ko, description_ko, video, audio_url, delivery_mode, unsupported_reason, translations]
      properties:
        sequence: { type: integer, minimum: 1 }
        task_code: { anyOf: [{ type: string, enum: [ONION_HARVEST, ONION_TRIMMING, ONION_SORTING, ONION_TRANSPORT, STRAWBERRY_HARVEST, STRAWBERRY_SORTING, STRAWBERRY_INSPECTION, STRAWBERRY_PACKING] }, { type: 'null' }] }
        title_ko: { type: string }
        description_ko: { type: string }
        video: { anyOf: [{ $ref: '#/components/schemas/VideoAsset' }, { type: 'null' }] }
        audio_url: { type: [string, 'null'], format: uri }
        delivery_mode: { type: string, enum: [VIDEO, TEXT_TTS, TEXT] }
        unsupported_reason: { type: [string, 'null'] }
        translations: { type: array, items: { $ref: '#/components/schemas/SegmentTranslation' } }
    VideoAsset:
      type: object
      required: [asset_id, task_code, video_url, provenance, review_status, safety_level, captions_text]
      properties:
        asset_id: { type: string }
        task_code: { type: string }
        video_url: { type: string, format: uri-reference }
        provenance: { type: string, const: AI_GENERATED_PREGENERATED }
        review_status: { type: string, enum: [PENDING, APPROVED, REJECTED] }
        safety_level: { type: string, enum: [LOW, HIGH] }
        captions_text: { type: string }
    SegmentTranslation:
      type: object
      required: [segment, language_code, text, source, verified, guide_lookup, phrase_key, source_page, source_url, license]
      properties:
        segment: { type: string, enum: [ACTION, QUANTITY, ORDER, SAFETY, LOCATION, OTHER] }
        language_code: { type: string, enum: [vi, ne] }
        text: { type: string }
        source: { type: string, enum: [OFFICIAL_GUIDE, AI_TRANSLATION, DETERMINISTIC] }
        verified: { type: boolean }
        guide_lookup: { type: string, enum: [HIT, MISS, NOT_APPLICABLE] }
        phrase_key: { type: [string, 'null'] }
        source_page: { type: [integer, 'null'] }
        source_url: { type: [string, 'null'] }
        license: { type: [string, 'null'] }
    WorkState:
      type: object
      required: [task_family, location, location_display, quantity, deadline, safety, notes, steps, risk_assessment, schema_version, contract_version, ontology_version]
      properties:
        task_family: { type: string, enum: [ONION, STRAWBERRY] }
        location: { $ref: '#/components/schemas/DeicticLocation' }
        location_display: { type: string }
        quantity: { anyOf: [{ $ref: '#/components/schemas/Quantity' }, { type: string, const: UNSPECIFIED }, { type: 'null' }] }
        deadline: { type: [string, 'null'] }
        safety: { type: array, items: { type: string } }
        notes: { type: [string, 'null'] }
        steps: { type: array, minItems: 0, items: { $ref: '#/components/schemas/Step' } }
        risk_assessment: { $ref: '#/components/schemas/RiskAssessment' }
        schema_version: { type: string, const: '2' }
        contract_version: { type: string, const: structure-v2 }
        ontology_version: { type: string, const: ontology-v2 }
...
    WorkVersion:
      type: object
      required: [version, lifecycle, state]
      properties:
        version: { type: integer, minimum: 1 }
        lifecycle: { type: string, enum: [PUBLISHED, SUPERSEDED] }
        state: { $ref: '#/components/schemas/WorkState' }
        ambiguity_override: { type: boolean }
        override_reason: { type: [string, 'null'], enum: [EXPERIENCED_WORKER, IN_PERSON_BRIEFING, OWNER_ACCEPTED_OTHER, null] }
        overridden_at: { type: [string, 'null'], format: date-time }
        transcript: { type: [string, 'null'] }
    OwnerWorkSession:
      type: object
      required: [session_id, current_version, lifecycle, version, contract_version, ontology_version]
      properties:
        session_id: { type: string }
        current_version: { type: integer, minimum: 1 }
        lifecycle: { type: string, const: PUBLISHED }
        version: { $ref: '#/components/schemas/WorkVersion' }
        contract_version: { type: string, const: structure-v2 }
        ontology_version: { type: string, const: ontology-v2 }
        worker_link_meta: { type: array, items: { $ref: '#/components/schemas/WorkerLinkMeta' } }
    WorkerLinkMeta:
      type: object
      required: [language_code, expires_at, revoked_at]
      properties:
        language_code: { type: string, enum: [vi, ne] }
        expires_at: { type: string, format: date-time }
        revoked_at: { type: [string, 'null'], format: date-time }
...
    LegacyWorkSessionRead:
      type: object
      additionalProperties: false
      required: [session_id, current_version, lifecycle, version, contract_version, ontology_version]
      properties:
        session_id: { type: string }
        current_version: { type: integer, minimum: 1 }
        lifecycle: { type: string, const: PUBLISHED }
        version: { $ref: '#/components/schemas/LegacyWorkVersion' }
        contract_version: { type: string, const: structure-v1 }
        ontology_version: { type: string, const: ontology-v1 }
        worker_link_meta: { type: array, items: { $ref: '#/components/schemas/WorkerLinkMeta' } }
    LegacyWorkVersion:
      type: object
      required: [version, lifecycle, state]
      properties:
        version: { type: integer, minimum: 1 }
        lifecycle: { type: string, enum: [PUBLISHED, SUPERSEDED] }
        state: { $ref: '#/components/schemas/LegacyWorkState' }
        ambiguity_override: { type: boolean }
        override_reason: { type: [string, 'null'], enum: [EXPERIENCED_WORKER, IN_PERSON_BRIEFING, OWNER_ACCEPTED_OTHER, null] }
        overridden_at: { type: [string, 'null'], format: date-time }
        transcript: { type: [string, 'null'] }
    LegacyWorkState:
      type: object
      required: [task_family, location, location_display, quantity, deadline, safety, notes, steps, risk_assessment, schema_version, contract_version, ontology_version]
      properties:
        task_family: { type: string, enum: [ONION, STRAWBERRY] }
        location: { $ref: '#/components/schemas/DeicticLocation' }
        location_display: { type: string }
        quantity: { anyOf: [{ $ref: '#/components/schemas/Quantity' }, { type: string, const: UNSPECIFIED }, { type: 'null' }] }
        deadline: { type: [string, 'null'] }
        safety: { type: array, items: { type: string } }
        notes: { type: [string, 'null'] }
        steps: { type: array, minItems: 0, items: { $ref: '#/components/schemas/LegacyStep' } }
        risk_assessment: { $ref: '#/components/schemas/RiskAssessment' }
        schema_version: { type: string }
        contract_version: { type: string }
        ontology_version: { type: string }
    LegacyStep:
      type: object
      required: [sequence, task_code, title_ko, description_ko, video, audio_url, delivery_mode, unsupported_reason, translations]
      properties:
        sequence: { type: integer, minimum: 1 }
        task_code: { type: [string, 'null'], description: Stored legacy codes are returned without remapping. }
        title_ko: { type: string }
        description_ko: { type: string }
        video: { anyOf: [{ $ref: '#/components/schemas/VideoAsset' }, { type: 'null' }] }
        audio_url: { type: [string, 'null'], format: uri }
        delivery_mode: { type: string, enum: [VIDEO, TEXT_TTS, TEXT] }
        unsupported_reason: { type: [string, 'null'] }
        translations: { type: array, items: { $ref: '#/components/schemas/SegmentTranslation' } }
...
    OwnerWorkSessionList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            oneOf:
              - { $ref: '#/components/schemas/OwnerWorkSession' }
              - { $ref: '#/components/schemas/LegacyWorkSessionRead' }
...
    WorkerBriefing:
      description: >-
        Locale-only vi/ne worker delivery DTO. It includes localized safety
        notices with verified provenance and omits transcript, risk assessment,
        identity/audit fields, token/cache metadata, and audio bytes. The
        opaque `tts.text_hash` is never displayed by worker UI.
      $ref: './schemas/worker-briefing-v2.schema.json'